Separate cli tests from gcloud tests
